Renames script-session -> session-script for consistency
