+++ src/main/webapp/scripts/session-script.js
