+++ src/main/webapp/session.html
@@ -3,7 +3,7 @@
   <meta charset="UTF-8"/>
   <title>VMN</title>
   <link rel="stylesheet" href="styles/session-style.css"/>
-  <script src="scripts/script-session.js"></script>
+  <script src="scripts/session-script.js"></script>
   <div id="header-bar">
     <h1 class="center-text">Virtual Movie Night</h1>
     <span
